Add validations in Json and Protobuf related to purity and nested serde protocol support

+++ krystal-common/src/main/java/com/flipkart/krystal/model/array/DoubleArray.java
@@ -3,7 +3,7 @@ package com.flipkart.krystal.model.array;
 import java.util.function.DoubleConsumer;
 import java.util.stream.DoubleStream;
 
-public non-sealed interface DoubleArray extends PrimitiveArray<Integer> {
+public non-sealed interface DoubleArray extends PrimitiveArray<Double> {
   double valueAt(int index);
 
   boolean contains(double target);
@@ -16,7 +16,7 @@ public non-sealed interface DoubleArray extends PrimitiveArray<Integer> {
 
   DoubleStream stream();
 
-  long[] toArray();
+  double[] toArray();
 
   DoubleArray subArray(int startIndexInclusive, int endIndexExclusive);
 }
